app/backend/requirements.txt
fastapi[standard]>=0.68.0
uvicorn>=0.15.0
pymongo==4.13.0
python-multipart>=0.0.5
numpy>=1.21.0
Pillow>=9.0.0
tritonclient[http]>=2.0.0 
opencv-python-headless
scipy